scripts/requirements.txt

google-generativeai
Pillow
